Fix test issues with ar in deterministic mode
The ar program in the binutils package in Debian is now configured
with --enable-deterministic-archives. This change makes the archives
reproducible, by setting the UID, GID, and timestamp to 0. However,
when dealing with archives created with the libxx(*.o) style rules,
make needs the timestamp of the file in order to decide to update it
or not. With the current deterministic behavior of ar, the timestamp
is always 0 and make fails to build, failing 7 out of 10 archive
tests.Since make seems to depend on timestamps for this rule it makes
sense to always pass on the -U flag for the test suite

This commit makes the minimal changes to make the test suite pass
whether or not ar defaults to using deterministic mode.

There is some online discussion:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=798804
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=798913
https://bugzilla.redhat.com/show_bug.cgi?id=1195883

From Manoj Srivastava <srivasta@debian.org>