Check for compiler flags and use GNU grep
In configure.ac check for valid compiler CFLAGS instead of them
hardcoded in `maintMakefile`. Still not perfect since Apple for example
to not return error for flags its compiler doesn't support.

Look for GNU grep instead of the system one. The `po-check` target uses
word boundary metacharacter `\b`. Asking the internet is seems as if
there are no overlapping word boundary metacharacters between the POSIX
and GNU grep implementations. And the stock OpenBSD is not GNU
compatible.