Change description of --tasks
I have come to the opinion `--tasks` should be
simpler in how it works.

Before, `remake --tasks` did some magic to see if a target was
"interesting", and one of the criteria was if it had a description
associated with it.

After years of using, I believe `--task`s should just look for targets
with descriptions. If that encourages people to document targets,
all the better.

Overall in this release I have elevated making use of target descriptions.

When we show a target now using the `target` command, we show any
description it has. And there is a `info tasks` which shows all targets
with descriptions.

Again, in my own experience here and in other projects, adding comments
and documentation is a good thing (tm); I think `remake` should be
doing whatever it can to promote this activity.

If limiting the checking that `--tasks` has to do to find its list of
targets nudges people into using the description comment more often,
then that in my opinion is a good thing.