Convert "chdir", "comment", and "info"
We are also doxygen commenting as we go along.

Note that as we convert .h's to .c's we must add specific header
dependencies for the commands. And commands like "set.h" which relied on
headers from prior commands need to be made explicit.

All of this is probably a good thing.