Added --profile-directory argument to specify profiling output location
If not used, then output is in the same directory that make is run
from. If provided and a path starting with / it will keep the
provided path. If it starts with a different character, it is
assumed to be a relative path, and is appended to the initial
starting directory