Improve location display with pattern rules
Consider the following makefile:

    %1: f2
        cp -T $< $@

    f2:
        touch f2

Before this patch, invoking the debugger with `remake -X f1` and then
running `backtrace` inside the debugger would display something like
this:

    remake<0> backtrace
    =>#0  f2 at /tmp/Makefile:4
    #1  f1 at ??

The location of f1 was not indicated and question marks were printed
instead.

A real-world situation where this happens is with Linux (at least for
version 5.18). After setting up a source tree with `make defconfig`,
`remake -X` displays question marks at the top of the backtrace.

This commit makes it so that when pattern rule search succeeds for a
target, its `floc` field is updated to the location of the pattern rule
that matches.

With this commit applied, the `backtrace` command mentioned above
displays this:

    remake<0> backtrace
    =>#0  f2 at /tmp/Makefile:4
    #1  f1 at /tmp/Makefile:1

A test case where the question marks were present in the expected output
is updated to match the new behavior.