Tweak MAKEFLAGS and -X interaction
Before this commit, running one of "continue", "next", "step" or "finish"
in the interactive debugger overwrote the value of MAKEFLAGS to either
"" or "-X".

This was sometimes a problem. For instance, with Linux v5.18, first
setting up the build system with `make defconfig`, launching the
debugger with `remake -X modules`, and running `continue` triggered
the following error:

	m2c    -o fs/efivarfs/efivarfs.o fs/efivarfs/efivarfs.mod
	remake[1]: m2c: No such file or directory
	<builtin>*** [fs/efivarfs/efivarfs.o] error 127

This happened because the root makefile disables Make's built-in rules
by appending "-R" to MAKEFLAGS, and the other makefiles rely on these
rules being disabled. `continue` overwrote MAKEFLAGS, which caused
recursive make instances to have built-in rules enabled.

This commit fixes this issue by changing the way the debugger commands
communicate status to other instances. It still happens through MAKEFLAGS,
but the value is now generated using the same function upstream Make uses.